splitting up unix from windows logic
